ensure that tests can't shell out to our real `git` and `open` executables
Fake `git` and `open` executables are pushed to PATH in testing and error
out when invoked. This is because there are tests that shell out to `git init`
and `git clone`, or open the browser, and bugs and failures in these tests
would previously cause them to poke around the system uncontrolled.