replace two dots (`sha1..sha2`) with three in the `compare` command
GitHub doesn't support ranges with two dots, only three. With these two
being equivalent in case of linear history, it's generally safe to convert
the range pasted on the command-line to three dots. This is to avoid
confusion where GitHub would display a 404 page otherwise.

See "specifying ranges" in the `git-rev-parse` man page.

closes #80