HUGE project-centric refactoring to support GitHub Enterprise
Add your GH Enterprise host to the list of known hosts:

  $ git config --global --add hub.host my.git.org

Configure your username and token for this host:

  $ git config --global github."my.git.org".user mislav
  $ git config --global github."my.git.org".token 1234abcd

Now a lot of operations should work for repos that were cloned from "my.git.org".

The default host is still github.com but this can be affected with the
`GITHUB_HOST` environment variable.

  $ GITHUB_HOST=my.git.org git clone myproject

Closes #98