Add single project mode to pull-request
If working on a single project owned by a organization, pull requests are
usually opened from one branch to another in the same repository. The old pull-
request required a full `-h user/repo:branch` specification for every new pull
request. This commit adds a new `-s` flag (for Single) that, in the absense
of any more information, uses the base project's user as the default user
for the head project as well.

Changing github.user to the target organization does not work becuase it
invalidates the API key.

This commit adds a single unit test w/ implementation.