make fork work when existing fork is the same as main project
Before hub fork would fail if your user already had a repo with the
same name no matter if the project was related or not. Added a
is_fork_of method that checks the repo_info to see if the github url
matches.

Updated fork.feature test to match new behavior but could not run tests so
not sure if that is 100% correct.