WIP merge.feature: Rework pull-request merges for the merge-button
I'm not sure if the second (private) merge is supposed to succeed or
not.  I'm also not sure if `hub merge …pull/…` should update the local
branch to reflect the changes (making it more like `git merge`) or if
the user should do that with a traditional Git fetch/merge (making it
more like the GitHub Merge Button).