make browser tests cross-platform
It was presumed that browse tests will take place on darwin only, as the
fakebin would catch the `open` command only. On machines that other than
darwin, web pages were actually openned during tests, but since they
were not openned via the fakebin, those scenarios were all counted as a
failure.

To prevent that, override the path of machine's default browser with a
symlink pointing to the fakebin. Then browse tests should be run
successfully, except for those platforms that do not support symlink...