pull-request: detect default branch for project
This is usually "master" but for e.g. git-flow projects it's "develop".
GitHub has an API field for the default branch but I avoided having to
hit the API and instead read from the "head" of the origin remote.

  git rev-parse --symbolic-full-name origin
  #=> refs/remotes/origin/develop

The "head" of the remote seems to be set at clone time by default, and
can be manually set with:

  git remote set-head origin develop

The pull-request command now detects this and sets the default branch as
the pull request base instead of always assuming "master" at origin.

Fixes #154, closes #326