Account for various locations of bash/zsh completion scripts
On different systems, bash/zsh completion scripts for git are installed
in various places. Account for that by searching the list of known paths.

In zsh mode, if a git-distributed zsh completion script can't be found
(e.g. on Ubuntu), ignore that and let it fall back to zsh-distributed
completion script.