Include head SHA in version info if tags are not available
When building hub from a git repository, the detailed version
information will get built into the executable courtesy of
`git describe --tags`. However, if tags are not available in the
git repo--for instance, in the case of Homebrew install from git--this
at least embeds the current SHA in version information.