Avoid shell redirection to silence stderr when shelling out
Shelling out to `git`, which by itself takes ~6ms, used to consistently
take 8-11ms in hub. Due to the number of different data we have to
collect from git, this overhead is the leading cause of slow execution.

Turns out, when we don't use shell redirection, but silence STDERR
ourselves in Ruby, this overhead is gone and the time spent shelling out
is much closer to git execution time.