Fork and set up remote in a non-blocking fashion
Forking through GitHub's API is asynchronous so there is really no way
of knowning when you can access git objects without polling.
Alternatively, we can establish the state of the fork with its remote
tracking branches and tags by fetching from the "origin" remote and then
changing the remote url to the fork.