Use the repo name in the response for fetch url.
The repository name of the pull request head doesn't necessarily match
the name of the project's repository. This can happen when a user forks
a project but already has a repository of the same name. Use the name
returned in the api response for the fetch url to avoid this edge case.