Don't re-wrap default pull request message taken from commit
Re-wrapping to 78 chars was dangerous since it could lead to moving
a #-reference to the beginning of a line which would then get
interpreted as a comment.

Wrapping pull request message is not useful anyway since it's Markdown.
For best results when rendered, users should unwrap their text.

Fixes #548, closes #570