Fix shebang on Ubuntu
Ubuntu is a kind of system where everything after the first space in a
shebang is provided as the first argument. Therefore, we can't allow
generating a script that has more than one argument to Ruby interpreter
in the shebang.

http://stackoverflow.com/a/4304187/11687

References ea0103fbc5ea61fc02577cf3ff3bdadda871dc20

Fixes #569