Add example for issue number
Technically, anything in the third argument is passes as a string, so you can look at any file if you know the hash (for instance, `blob/HASH/app/index.js`. However, having the issue number passed seems to me to be something really useful to my workflow, so I was hoping to find something in the documentation about it. 

Another possibility would be to enable a fourth argument for issue number. I'm happy to do a PR for this if something things it might be of more worth than just this edit.