Stop trying to re-use existing Access Tokens for hub
Due to upcoming GitHub API changes, the "token" field will not be
accessible anymore, which will make it impossible for hub to re-use an
existing token that might have been generated for another computer.

Now, a new Access Token will be created every time hub needs to exchange
the username/password credentials for an OAuth token.

https://developer.github.com/changes/2014-12-08-removing-authorizations-token/