Don't suggest piping `pull-request | pbcopy` because of vim
With an interactive editor popping up during `pull-request`, it's not
advisable to redirect stdout to somewhere other than the terminal
because vim will complain. However, use this example to demonstrate how
you can skip the editor if you've already prepared the pull request
message in a file (in a corporate environment, likely with some
automated scripts).