Properly support `-C` git global flag
From git documentation:

    -C <path> : Run as if git was started in <path> instead of the
                current working directory. When multiple -C options are
                given, each subsequent non-absolute -C <path> is
                interpreted relative to the preceding -C <path>.

We correctly forwarded the `-C <path>` values as git global flags to git
invocations such as `git rev-parse --git-dir`, but that command is
designed to return the result as a relative path to the value computed
from `-C`. This adds an extra step to transform this relative path to an
absolute one by applying values of `-C` as the base directory.