Added `upstream` option to `browse`
I constantly want to be able to go directly to the remote upstream webpage for
the repo I am currently working on. This normally enters my workflow when I am
working on a fork, push to a feature branch, and want to PR to the remote repo
upstream. I would run `gh browse`, then click on the upstream link on the website,
then start the PR process. Having `gh browse upstream` will make this much easier.

This can be extended to deal with other remotes as well, perhaps using a subpage
like system: `git browse remote upstream`. I am happy to do this change, as well,
but figured a smaller fix like this would open discussion easier.