insert vim config at front of command
when using mvim as your EDITOR, the option `--remote-wait` must be 
supplied to ensure that the `hub` command does not continue until the 
file has been edited and closed.

since the `--remote-wait` option treats all following arguments as a 
file list, the hub-specific options passed to vim *must* be added before 
the `--remote-wait` option. if they are added after, vim tries to edit 
a file named `-c`, which is definitely not what anyone wants.