If found, reuse existing git remote for PR head in `checkout`
When there is already a git remote for the head repo of a PR to be
checked out, fetch the corresponding branch from that remote instead of
fetching the special `refs/pull/XY/head` ref.

This sets up a pull/push workflow for branches that the user might have
write access to and wishes to collaborate on.

Fixes #1238