Remove unhelpful test
With "hub checkout" you had to write the complete URL to the PR, which
made it possible to type the wrong owner/repo there. This check covers
that case.

Nowadays, when you only enter the PR number you can't mess up that way
any more, and this test lacks meaning.