Make label a sub-command of issue
According to the discussion on #1545 we want to have as few top-level
hub commands as possible to reduce the chances of colliding with
exisitng aliases or commands which may be added to git in the future.

Also, the label listing and modification functionality is not likely
to be used very often, so we shouldn't give the set of commands for
interacting with labels the status of a top-level command.