Ensure consistent sort direction when listing issues, PRs
The default direction when sorting by `created` date (the default) is
descending. However, the default on the API backend changes to ascending
when the sort key is something different than `created`:
https://developer.github.com/v3/pulls/#list-pull-requests

This ensures that the default `direction: "desc"` is always passed to
API queries, regardless of the sort key, unless `--sort-ascending` was used.

Ref. https://github.com/github/hub/issues/1815#issuecomment-433222344