Pass $EDITOR to the shell
Several attempts have been made to provide better handling of the EDITOR
and VISUAL environment variables. However, because these commands can
contain arbitrary shell code, it isn't possible to provide correct
behavior without actually invoking the shell.

Create a new function that invokes the shell with the proper arguments.
For all editor invocations, invoke the shell in the way that Git does,
even though this is incompatible with the way that other software
behaves. For example, Git allows the use of "$@" directly in the
command, while most other programs, including Debian's sensible-editor
and Git LFS, require the user to wrap such commands in a shell function.
It isn't possible to make it work in both cases, but we're trying to be
like Git, so do what Git does.

Rewrite the command line in verbose output so we don't show the user the
doubled argument we need to pass for things to work correctly. Also,
stop expanding environment variables in the editor code path, since the
shell will do that for us.