Don't suggest completions for git fork free-text arguments
The --remote-name and --org parameters to git fork expect arguments
which can be considered free text because the user has to suggest a
name, so we can't provide suggestions there.

If one of those two was the previous word then we clear completion
suggestions until a free-text word has been entered, allowing the user
to add their own input there.