Make --no-remote and --remote-name completions mutually exclusive
Because it wouldn't make sense to set a name for a remote when you ask
it not to add a remote (or vice versa) these two flags are mutually
exclusive.

For lack of an idea for a more elegant solution, I've made it so that
whenever either of them matches, the other will also be dropped from the
available flags list.  This works perfectly for now but will probably
need to be revisited if any other such parameters are added in future.