Remove URL output from successful `issue update`
This made sense for `issue create`, where a person would want to obtain
the URL to the new issue, but it makes less sense for `issue update`,
where the user already knows the issue number (which will not change).

This makes it so that the successful run has no output, but since I
don't know what a better confirmation behavior would be, I think this is
fine for now.