Add format placeholder for asset URLs and labels
For automated deployment it is useful to get the asset URL after creating a release, but with the existing asset list placeholder (%as) we get both the URL and its label, separated by a horizontal tab character, this is even when the label is empty, requiring us to post-process the output from `hub release show --format %as <TAG>`.

This commit introduces a new placeholder (%au) which allows us to obtain the URL without having to do any parsing.

For consistency we also add a placeholder for asset labels (%al).