bash completion: add support for git 2.30 and on
The _git backwards compat wrapper was dropped upstream:
https://github.com/git/git/commit/441ecdab37fefdacf32575a60aa523b2367c46f7

Instead, we rely on __git_complete to detect loaded bash completions,
since we will use it later on to set up completions -- it is now public
API.

There is a gap between git 2.30 and 2.31 where _git does not exist, but
there is no public API to create completions. Starting 2.31, we are
formally permitted to copy/imitate the upstream completions with:

```
__git_complete mycmd git_cmd
```

For 2.30 specifically, we have to pass the internal completion function
used instead of "git_cmd", but it's difficult to detect this necessity
in a forwards-compatible way. Try it that way first, in the assumption
that if the internal completion function still exists it does the same
thing.